+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## v0.8.3 - 2024-10-18
 
-New `if` option for replicas which decides if they should be deployed or not.
+New `if` option for replicas which decides if they should be updated or not.
 
-- **If not provided**, the replica will be deployed (so no impact on existing configurations)
+- **If not provided**, the replica will be updated (so no impact on existing configurations)
 - Must be `nil|true|false` or the name (atom) of a arity-0 func which returns a boolean
-- If provided, the replica will be deployed only if the value is `true` or the function returns `true`
+- If provided, the replica will be updated only if the value is `true` or the function returns `true`
 
 Useful for Algolia's A/B testing which requires replicas and only deploy them in production.